clean many files
